+++ install.py
@@ -36,7 +36,7 @@ freertos_sdk_path = install_path
 print("Standalone SDK at {}".format(freertos_sdk_path))
 
 # Add standalone sdk
-standalone_sdk_v="14b5871f36fe95638c390c81a959641f62bee761"
+standalone_sdk_v="cd15b23000d0c52968dc322b6d99025195bbf446"
 if (install_platform == windows_x64):
     standalone_path=freertos_sdk_path  + '\\standalone'
 else: